Further refractor of the file structure
