+++ app/models/machine.py
