+++ tests/constants.py
@@ -1,0 +1,27 @@
+"""
+This file holds all the constants used in the test files.
+"""
+
+# Test Parameters for User
+USER_TEST_ID = 1
+USER_TEST_NAME = "Hayden"
+USER_TEST_PUBLIC_ID = "10101"
+USER_TEST_EMAIL = "Walla Walla"
+USER_TEST_FLOOR = 2
+USER_TEST_DORM = "Foreman"
+
+# Test Parameters for Machine
+MACHINE_TEST_PUBLIC_ID = 1
+MACHINE_TEST_FLOOR_ID = 1
+MACHINE_TEST_DORM = "Sittner"
+MACHINE_TEST_FLOOR = 1
+MACHINE_TEST_STATUS = "Free"
+MACHINE_TEST_LAST_SERVICE_DATE = "10-27-2022"
+MACHINE_TEST_INSTALLATION_DATE = "10-27-2022"
+MACHINE_TEST_FINISH_TIME = None
+MACHINE_TEST_USER_NAME = None
+
+# Test Parameters for Emails
+RECIPIENTS = ["WWU-Wash-And-Dry@outlook.com"]
+BODY = "Testing email"
+SUBJECT = "Testing"
